change sort by relevance test to test `SearchResultListSorter`
